+++ packages/terra-avatar/src/terra-dev-site/doc/example/generic/provider/Provider.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Generic } from 'terra-avatar';
+
+const AvatarProvider = () => (
+  <Generic variant="provider" alt="Provider" />
+);
+
+export default AvatarProvider;