+++ packages/terra-avatar/src/terra-dev-site/doc/example/generic/provider/ProviderSize.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Generic } from 'terra-avatar';
+
+const AvatarSize = () => (
+  <Generic variant="provider" alt="Provider" size="2em" />
+);
+
+export default AvatarSize;