+++ packages/terra-avatar/src/terra-dev-site/doc/example/generic/shared-user/GenericSharedUserSize.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Generic } from 'terra-avatar';
+
+const AvatarSize = () => (
+  <Generic variant="shared-user" alt="Shared User" size="2em" />
+);
+
+export default AvatarSize;