+++ packages/terra-avatar/src/terra-dev-site/doc/example/generic/single-user/GenericSingleUser.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Generic } from 'terra-avatar';
+
+const AvatarSingleUser = () => (
+  <Generic variant="single-user" alt="Single User" />
+);
+
+export default AvatarSingleUser;