+++ packages/terra-avatar/src/terra-dev-site/doc/example/generic/single-user/GenericSingleUserSize.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Generic } from 'terra-avatar';
+
+const AvatarSize = () => (
+  <Generic variant="single-user" alt="Single User" size="2em" />
+);
+
+export default AvatarSize;