+++ packages/terra-avatar/src/terra-dev-site/test/avatar/generic/SharedUser.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { Generic } from '../../../../index';
+
+export default () => <Generic id="shared-user" variant="shared-user" alt="shared user" color="eight" />;