+++ packages/terra-avatar/src/terra-dev-site/test/avatar/generic/SingleUser.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { Generic } from '../../../../index';
+
+export default () => <Generic id="single-user" alt="single user" color="four" />;