+++ packages/terra-avatar/src/terra-dev-site/test/avatar/shared-user/DefaultSharedUser.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import { SharedUser } from '../../../../index';
-
-export default () => <SharedUser id="shared-user" alt="shared" color="four" />;